+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## 0.61.0 (2022/06/06)
 
-* [ENHANCEMENT] Added for cloud LB support. [237](https://github.com/ortuman/jackal/pull/237) 
+* [ENHANCEMENT] Helm: added support for cloud LB. [237](https://github.com/ortuman/jackal/pull/237) 
 
 ## 0.60.0 (2022/05/27)