+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## jackal - main / unreleased
 
 * [ENHANCEMENT] Re-enable TLS 1.3 channel binding during auth using [RFC 9266](https://www.rfc-editor.org/rfc/rfc9266).
+* [ENHANCEMENT] hook: include propagated context into execution parameter. [249](https://github.com/ortuman/jackal/pull/249)
 
 ## 0.61.0 (2022/06/06)